% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{set_intersection}
\alias{set_intersection}
\title{Get the intersection between sets}
\usage{
set_intersection(x, y, ...)
}
\description{
Get the values that intersection between x and all containers in
...
}
