% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_difference.R
\name{set_symmetric_difference}
\alias{set_symmetric_difference}
\title{Get the symmetric difference between sets}
\usage{
set_symmetric_difference(x, y)
}
\description{
Return a new set with elements in either set but not in both.
}
