\name{sets-internal}
\title{Internal}
\alias{.cartesian_product}
\alias{.exact_match}
\alias{.get_memberships}
\alias{.make_gset_from_support_and_memberships}
\alias{.make_set_of_tuples_from_list_of_lists}
\alias{.ntuple}
\alias{.set_subset}
\alias{fuzzy_logic_predicates}
\description{Internal functions not intended for public use.}
\keyword{ internal }
