% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{sets_difference}
\alias{sets_difference}
\title{Compute the Difference Between Sets}
\usage{
sets_difference(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional vector sets to compare against \code{x}.}
}
\value{
A vector containing elements unique to \code{x}.
}
\description{
\code{set_difference} returns the elements in \code{x} that are not in the other sets.
This function mimics the behavior of Python's \code{set.difference()} method.
}
\examples{
sets_difference(c(1, 2, 3), c(2, 3, 4))
}
\seealso{
\href{https://docs.python.org/3/library/stdtypes.html#frozenset.difference}{Python set.difference() documentation}
}
