% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal.R
\name{sets_equal}
\alias{sets_equal}
\title{Check if Two Sets are Equal}
\usage{
sets_equal(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{Another vector to compare with \code{x}.}
}
\value{
\code{TRUE} if \code{x} and \code{y} contain the same elements, otherwise \code{FALSE}.
}
\description{
\code{set_equal} checks if two sets \code{x} and \code{y} contain the same elements.
This function is similar to Python's \code{set.__eq__()} method.
}
\examples{
sets_equal(c(1, 2, 3), c(3, 1, 2))
}
