% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{sets_intersection}
\alias{sets_intersection}
\title{Compute the Intersection of Sets}
\usage{
sets_intersection(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional vectors to find the intersection with.}
}
\value{
A vector containing elements common to all sets.
}
\description{
\code{set_intersection} returns the common elements in all sets.
This function mimics Python's \code{set.intersection()} method.
}
\examples{
set_intersection(c(1, 2, 3), c(2, 3, 4), c(3, 4, 5))
}
\seealso{
\href{https://docs.python.org/3/library/stdtypes.html#frozenset.intersection}{Python set.intersection() documentation}
}
