% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issubset.R
\name{sets_issubset}
\alias{sets_issubset}
\title{Check if a Set is a Subset of Another Set}
\usage{
sets_issubset(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{Another vector to compare with.}
}
\value{
\code{TRUE} if \code{x} is a subset of \code{y}, otherwise \code{FALSE}.
}
\description{
\code{set_issubset} checks if all elements of \code{x} are in \code{y}.
This function is similar to Python's \code{set.issubset()} method.
}
\examples{
sets_issubset(c(1, 2), c(1, 2, 3))
}
\seealso{
\href{https://docs.python.org/3/library/stdtypes.html#frozenset.issubset}{Python set.issubset() documentation}
}
