% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuperset.R
\name{sets_issuperset}
\alias{sets_issuperset}
\title{Check if a Set is a Superset of Another Set}
\usage{
sets_issuperset(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{Another vector to compare with.}
}
\value{
\code{TRUE} if \code{x} is a superset of \code{y}, otherwise \code{FALSE}.
}
\description{
\code{set_issuperset} checks if all elements of \code{y} are in \code{x}.
This function is similar to Python's \code{set.issuperset()} method.
}
\examples{
sets_issuperset(c(1, 2, 3), c(2, 3))
}
\seealso{
\href{https://docs.python.org/3/library/stdtypes.html#frozenset.issuperset}{Python set.issuperset() documentation}
}
