% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_difference.R
\name{sets_symmetric_difference}
\alias{sets_symmetric_difference}
\title{Compute the Symmetric Difference Between Sets}
\usage{
sets_symmetric_difference(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{Another vector to compare with.}
}
\value{
A vector containing elements unique to either \code{x} or \code{y}.
}
\description{
\code{set_symmetric_difference} returns elements in either \code{x} or \code{y} but not in
both. This function mimics Python's \code{set.symmetric_difference()} method.
}
\examples{
sets_symmetric_difference(c(1, 2, 3), c(3, 4, 5))
}
\seealso{
\href{https://docs.python.org/3/library/stdtypes.html#frozenset.symmetric_difference}{Python set.symmetric_difference() documentation} # nolint
}
