% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{sets_union}
\alias{sets_union}
\title{Compute the Union of Sets}
\usage{
sets_union(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional vectors to unite with \code{x}.}
}
\value{
A set containing all unique elements from all vectors.
}
\description{
\code{set_union} returns all elements present in any of the sets.
This function is similar to Python's \code{set.union()} method.
}
\examples{
set_union(c(1, 2, 3), c(3, 4, 5))
}
\seealso{
\href{https://docs.python.org/3/library/stdtypes.html#frozenset.union}{Python set.union() documentation}
}
